Create Sum of Prime numbers using Sieve of Eratosthenes
Sieve of Eratosthenes: This algorithm efficiently finds all prime numbers up to a given limit by iteratively marking the multiples of each prime number as composite (not prime). Here's how it works in the code:

You are given a positive integer N, return the sum of all prime numbers between 1 and N(inclusive).